requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic-settings==2.1.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.23.2

